Restructured software architecture - Needed common library folder for make files - Moved test cases and old make files in to legacy-code - removed lib folder to merge in to common (easier to integrate architecture) - started makefiles
