--- lib/Utilities/.gitignore
@@ -1,3 +1,0 @@
-#placeholder
-*
-!.gitignore
